refactor: split more api clients
