+++ src/ui/main-axios.ts
@@ -12,10 +12,6 @@ import type {
   C2STunnelPreset,
   FileManagerFile,
   FileManagerShortcut,
-  DockerContainer,
-  DockerStats,
-  DockerLogOptions,
-  DockerValidation,
   ProxyNode,
 } from "@/types/index";
 
@@ -3646,758 +3642,62 @@ export async function unlinkOIDCFromPasswordAccount(
   }
 }
 
-export interface GuacamoleTokenRequest {
-  protocol: "rdp" | "vnc" | "telnet";
-  hostname: string;
-  port?: number;
-  username?: string;
-  password?: string;
-  domain?: string;
-  security?: string;
-  ignoreCert?: boolean;
-  guacamoleConfig?: {
-    colorDepth?: number;
-    width?: number;
-    height?: number;
-    dpi?: number;
-    resizeMethod?: string;
-    forceLossless?: boolean;
-    disableAudio?: boolean;
-    enableAudioInput?: boolean;
-    enableWallpaper?: boolean;
-    enableTheming?: boolean;
-    enableFontSmoothing?: boolean;
-    enableFullWindowDrag?: boolean;
-    enableDesktopComposition?: boolean;
-    enableMenuAnimations?: boolean;
-    disableBitmapCaching?: boolean;
-    disableOffscreenCaching?: boolean;
-    disableGlyphCaching?: boolean;
-    disableGfx?: boolean;
-    enablePrinting?: boolean;
-    printerName?: string;
-    enableDrive?: boolean;
-    driveName?: string;
-    drivePath?: string;
-    createDrivePath?: boolean;
-    disableDownload?: boolean;
-    disableUpload?: boolean;
-    enableTouch?: boolean;
-    clientName?: string;
-    console?: boolean;
-    initialProgram?: string;
-    serverLayout?: string;
-    timezone?: string;
-    gatewayHostname?: string;
-    gatewayPort?: number;
-    gatewayUsername?: string;
-    gatewayPassword?: string;
-    gatewayDomain?: string;
-    remoteApp?: string;
-    remoteAppDir?: string;
-    remoteAppArgs?: string;
-    normalizeClipboard?: string;
-    disableCopy?: boolean;
-    disablePaste?: boolean;
-    cursor?: string;
-    swapRedBlue?: boolean;
-    readOnly?: boolean;
-    recordingPath?: string;
-    recordingName?: string;
-    createRecordingPath?: boolean;
-    recordingExcludeOutput?: boolean;
-    recordingExcludeMouse?: boolean;
-    recordingIncludeKeys?: boolean;
-    wolSendPacket?: boolean;
-    wolMacAddr?: string;
-    wolBroadcastAddr?: string;
-    wolUdpPort?: number;
-    wolWaitTime?: number;
-  };
-}
-
-export interface GuacamoleTokenResponse {
-  token: string;
-}
-
-type GuacamoleConfigSource = {
-  guacamoleConfig?: string | Record<string, unknown> | null;
-};
-
-export function getGuacamoleDpi(
-  source?: GuacamoleConfigSource,
-): number | undefined {
-  const config = source?.guacamoleConfig;
-  if (!config) return undefined;
-
-  let dpi: unknown;
-  if (typeof config === "string") {
-    try {
-      dpi = JSON.parse(config).dpi;
-    } catch {
-      return undefined;
-    }
-  } else {
-    dpi = config.dpi;
-  }
-
-  const parsedDpi = typeof dpi === "string" ? Number(dpi) : dpi;
-  if (
-    typeof parsedDpi !== "number" ||
-    !Number.isFinite(parsedDpi) ||
-    parsedDpi <= 0
-  ) {
-    return undefined;
-  }
-
-  return Math.trunc(parsedDpi);
-}
-
-function toGuacamoleParams(
-  config: GuacamoleTokenRequest["guacamoleConfig"],
-): Record<string, unknown> {
-  if (!config) return {};
-
-  const params: Record<string, unknown> = {};
-
-  const mappings: Record<string, string> = {
-    colorDepth: "color-depth",
-    resizeMethod: "resize-method",
-    forceLossless: "force-lossless",
-    disableAudio: "disable-audio",
-    enableAudioInput: "enable-audio-input",
-    enableWallpaper: "enable-wallpaper",
-    enableTheming: "enable-theming",
-    enableFontSmoothing: "enable-font-smoothing",
-    enableFullWindowDrag: "enable-full-window-drag",
-    enableDesktopComposition: "enable-desktop-composition",
-    enableMenuAnimations: "enable-menu-animations",
-    disableBitmapCaching: "disable-bitmap-caching",
-    disableOffscreenCaching: "disable-offscreen-caching",
-    disableGlyphCaching: "disable-glyph-caching",
-    disableGfx: "disable-gfx",
-    enablePrinting: "enable-printing",
-    printerName: "printer-name",
-    enableDrive: "enable-drive",
-    driveName: "drive-name",
-    drivePath: "drive-path",
-    createDrivePath: "create-drive-path",
-    disableDownload: "disable-download",
-    disableUpload: "disable-upload",
-    enableTouch: "enable-touch",
-    clientName: "client-name",
-    initialProgram: "initial-program",
-    serverLayout: "server-layout",
-    gatewayHostname: "gateway-hostname",
-    gatewayPort: "gateway-port",
-    gatewayUsername: "gateway-username",
-    gatewayPassword: "gateway-password",
-    gatewayDomain: "gateway-domain",
-    remoteApp: "remote-app",
-    remoteAppDir: "remote-app-dir",
-    remoteAppArgs: "remote-app-args",
-    normalizeClipboard: "normalize-clipboard",
-    disableCopy: "disable-copy",
-    disablePaste: "disable-paste",
-    swapRedBlue: "swap-red-blue",
-    readOnly: "read-only",
-    recordingPath: "recording-path",
-    recordingName: "recording-name",
-    createRecordingPath: "create-recording-path",
-    recordingExcludeOutput: "recording-exclude-output",
-    recordingExcludeMouse: "recording-exclude-mouse",
-    recordingIncludeKeys: "recording-include-keys",
-    wolSendPacket: "wol-send-packet",
-    wolMacAddr: "wol-mac-addr",
-    wolBroadcastAddr: "wol-broadcast-addr",
-    wolUdpPort: "wol-udp-port",
-    wolWaitTime: "wol-wait-time",
-  };
-
-  for (const [key, value] of Object.entries(config)) {
-    if (value !== undefined && value !== null && value !== "") {
-      const paramName = mappings[key] || key;
-      if (typeof value === "boolean") {
-        params[paramName] = value ? "true" : "false";
-      } else {
-        params[paramName] = value;
-      }
-    }
-  }
-
-  return params;
-}
-
-export async function getGuacamoleToken(
-  request: GuacamoleTokenRequest,
-): Promise<GuacamoleTokenResponse> {
-  try {
-    const guacParams = toGuacamoleParams(request.guacamoleConfig);
-
-    const response = await authApi.post("/guacamole/token", {
-      type: request.protocol,
-      hostname: request.hostname,
-      port: request.port,
-      username: request.username,
-      password: request.password,
-      domain: request.domain,
-      security: request.security,
-      "ignore-cert": request.ignoreCert,
-      ...guacParams,
-    });
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "get guacamole token");
-  }
-}
-
-export async function getGuacamoleTokenFromHost(
-  hostId: number,
-  protocol?: "rdp" | "vnc" | "telnet",
-): Promise<GuacamoleTokenResponse> {
-  try {
-    const response = await authApi.post(
-      `/guacamole/connect-host/${hostId}`,
-      protocol ? { protocol } : {},
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "get guacamole token from host");
-  }
-}
-
-export async function getGuacdStatus(): Promise<{
-  guacd: { status: string };
-}> {
-  const response = await authApi.get("/guacamole/status");
-  return response.data;
-}
+export type {
+  GuacamoleTokenRequest,
+  GuacamoleTokenResponse,
+} from "@/api/guacamole-api";
+export {
+  getGuacamoleDpi,
+  getGuacamoleToken,
+  getGuacamoleTokenFromHost,
+  getGuacdStatus,
+} from "@/api/guacamole-api";
 
 // ============================================================================
 // RBAC MANAGEMENT
 // ============================================================================
 
-export async function getRoles(): Promise<{ roles: Role[] }> {
-  try {
-    const response = await rbacApi.get("/rbac/roles");
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "fetch roles");
-  }
-}
-
-export async function createRole(roleData: {
-  name: string;
-  displayName: string;
-  description?: string | null;
-}): Promise<{ role: Role }> {
-  try {
-    const response = await rbacApi.post("/rbac/roles", roleData);
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "create role");
-  }
-}
-
-export async function updateRole(
-  roleId: number,
-  roleData: {
-    displayName?: string;
-    description?: string | null;
-  },
-): Promise<{ role: Role }> {
-  try {
-    const response = await rbacApi.put(`/rbac/roles/${roleId}`, roleData);
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "update role");
-  }
-}
-
-export async function deleteRole(
-  roleId: number,
-): Promise<{ success: boolean }> {
-  try {
-    const response = await rbacApi.delete(`/rbac/roles/${roleId}`);
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "delete role");
-  }
-}
-
-export async function getUserRoles(
-  userId: string,
-): Promise<{ roles: UserRole[] }> {
-  try {
-    const response = await rbacApi.get(`/rbac/users/${userId}/roles`);
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "fetch user roles");
-  }
-}
-
-export async function assignRoleToUser(
-  userId: string,
-  roleId: number,
-): Promise<{ success: boolean }> {
-  try {
-    const response = await rbacApi.post(`/rbac/users/${userId}/roles`, {
-      roleId,
-    });
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "assign role to user");
-  }
-}
-
-export async function removeRoleFromUser(
-  userId: string,
-  roleId: number,
-): Promise<{ success: boolean }> {
-  try {
-    const response = await rbacApi.delete(
-      `/rbac/users/${userId}/roles/${roleId}`,
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "remove role from user");
-  }
-}
-
-export async function shareHost(
-  hostId: number,
-  shareData: {
-    targetType: "user" | "role";
-    targetUserId?: string;
-    targetRoleId?: number;
-    permissionLevel: "view";
-    durationHours?: number;
-  },
-): Promise<{ success: boolean }> {
-  try {
-    const response = await rbacApi.post(
-      `/rbac/host/${hostId}/share`,
-      shareData,
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "share host");
-  }
-}
-
-export async function getHostAccess(
-  hostId: number,
-): Promise<{ accessList: AccessRecord[] }> {
-  try {
-    const response = await rbacApi.get(`/rbac/host/${hostId}/access`);
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "fetch host access");
-  }
-}
-
-export async function revokeHostAccess(
-  hostId: number,
-  accessId: number,
-): Promise<{ success: boolean }> {
-  try {
-    const response = await rbacApi.delete(
-      `/rbac/host/${hostId}/access/${accessId}`,
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "revoke host access");
-  }
-}
-
-// ============================================================================
-// SNIPPET SHARING
-// ============================================================================
-
-export async function shareSnippet(
-  snippetId: number,
-  shareData: {
-    targetType: "user" | "role";
-    targetUserId?: string;
-    targetRoleId?: number;
-    durationHours?: number;
-  },
-): Promise<{ success: boolean }> {
-  try {
-    const response = await rbacApi.post(
-      `/rbac/snippet/${snippetId}/share`,
-      shareData,
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "share snippet");
-  }
-}
-
-export async function getSnippetAccess(
-  snippetId: number,
-): Promise<{ accessList: AccessRecord[] }> {
-  try {
-    const response = await rbacApi.get(`/rbac/snippet/${snippetId}/access`);
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "fetch snippet access");
-  }
-}
-
-export async function revokeSnippetAccess(
-  snippetId: number,
-  accessId: number,
-): Promise<{ success: boolean }> {
-  try {
-    const response = await rbacApi.delete(
-      `/rbac/snippet/${snippetId}/access/${accessId}`,
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "revoke snippet access");
-  }
-}
-
-export async function getSharedSnippets(): Promise<{
-  sharedSnippets: Array<{
-    id: number;
-    name: string;
-    content: string;
-    description: string | null;
-    folder: string | null;
-    ownerUsername: string;
-    permissionLevel: string;
-    expiresAt: string | null;
-  }>;
-}> {
-  try {
-    const response = await rbacApi.get("/rbac/shared-snippets");
-    return response.data;
-  } catch (error) {
-    handleApiError(error, "fetch shared snippets");
-  }
-}
+export {
+  getRoles,
+  createRole,
+  updateRole,
+  deleteRole,
+  getUserRoles,
+  assignRoleToUser,
+  removeRoleFromUser,
+  shareHost,
+  getHostAccess,
+  revokeHostAccess,
+  shareSnippet,
+  getSnippetAccess,
+  revokeSnippetAccess,
+  getSharedSnippets,
+} from "@/api/rbac-api";
 
 // ============================================================================
 // DOCKER MANAGEMENT API
 // ============================================================================
 
-export async function connectDockerSession(
-  sessionId: string,
-  hostId: number,
-  config?: {
-    userProvidedPassword?: string;
-    userProvidedSshKey?: string;
-    userProvidedKeyPassword?: string;
-    forceKeyboardInteractive?: boolean;
-    useSocks5?: boolean;
-    socks5Host?: string;
-    socks5Port?: number;
-    socks5Username?: string;
-    socks5Password?: string;
-    socks5ProxyChain?: unknown;
-  },
-): Promise<{
-  success?: boolean;
-  message?: string;
-  requires_totp?: boolean;
-  prompt?: string;
-  isPassword?: boolean;
-  status?: string;
-  reason?: string;
-  connectionLogs?: ApiConnectionLog[];
-  requires_warpgate?: boolean;
-  url?: string;
-  securityKey?: string;
-}> {
-  try {
-    const response = await dockerApi.post("/ssh/connect", {
-      sessionId,
-      hostId,
-      ...config,
-    });
-    return response.data;
-  } catch (error: unknown) {
-    if (
-      axios.isAxiosError<ConnectErrorResponse>(error) &&
-      error.response?.data?.status === "auth_required"
-    ) {
-      return error.response.data;
-    }
-    if (
-      axios.isAxiosError<ConnectErrorResponse>(error) &&
-      error.response?.data?.requires_totp
-    ) {
-      return error.response.data;
-    }
-    if (
-      axios.isAxiosError<ConnectErrorResponse>(error) &&
-      error.response?.data?.requires_warpgate
-    ) {
-      return error.response.data;
-    }
-    if (
-      axios.isAxiosError<ConnectErrorResponse>(error) &&
-      error.response?.data?.connectionLogs
-    ) {
-      const data = error.response.data;
-      const errorWithLogs = new Error(
-        data.error || data.message || error.message,
-      );
-      Object.assign(errorWithLogs, {
-        connectionLogs: data.connectionLogs,
-      });
-      throw errorWithLogs;
-    }
-    throw handleApiError(error, "connect to Docker SSH session");
-  }
-}
-
-export async function verifyDockerTOTP(
-  sessionId: string,
-  totpCode: string,
-): Promise<{ status: string; message: string }> {
-  try {
-    const response = await dockerApi.post("/ssh/connect-totp", {
-      sessionId,
-      totpCode,
-    });
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "verify Docker TOTP");
-  }
-}
-
-export async function verifyDockerWarpgate(
-  sessionId: string,
-): Promise<{ status: string; message: string }> {
-  try {
-    const response = await dockerApi.post("/ssh/connect-warpgate", {
-      sessionId,
-    });
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "verify Docker Warpgate");
-  }
-}
-
-export async function disconnectDockerSession(
-  sessionId: string,
-): Promise<{ success: boolean; message: string }> {
-  try {
-    const response = await dockerApi.post("/ssh/disconnect", {
-      sessionId,
-    });
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "disconnect from Docker SSH session");
-  }
-}
-
-export async function keepaliveDockerSession(
-  sessionId: string,
-): Promise<{ success: boolean }> {
-  try {
-    const response = await dockerApi.post("/ssh/keepalive", {
-      sessionId,
-    });
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "keepalive Docker SSH session");
-  }
-}
-
-export async function getDockerSessionStatus(
-  sessionId: string,
-): Promise<{ success: boolean; connected: boolean }> {
-  try {
-    const response = await dockerApi.get("/ssh/status", {
-      params: { sessionId },
-    });
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "get Docker session status");
-  }
-}
-
-export async function validateDockerAvailability(
-  sessionId: string,
-): Promise<DockerValidation> {
-  try {
-    const response = await dockerApi.get(`/validate/${sessionId}`);
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "validate Docker availability");
-  }
-}
-
-export async function listDockerContainers(
-  sessionId: string,
-  all: boolean = true,
-): Promise<DockerContainer[]> {
-  try {
-    const response = await dockerApi.get(`/containers/${sessionId}`, {
-      params: { all },
-    });
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "list Docker containers");
-  }
-}
-
-export async function getDockerContainerDetails(
-  sessionId: string,
-  containerId: string,
-): Promise<DockerContainer> {
-  try {
-    const response = await dockerApi.get(
-      `/containers/${sessionId}/${containerId}`,
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "get Docker container details");
-  }
-}
-
-export async function startDockerContainer(
-  sessionId: string,
-  containerId: string,
-): Promise<{ success: boolean; message: string }> {
-  try {
-    const response = await dockerApi.post(
-      `/containers/${sessionId}/${containerId}/start`,
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "start Docker container");
-  }
-}
-
-export async function stopDockerContainer(
-  sessionId: string,
-  containerId: string,
-): Promise<{ success: boolean; message: string }> {
-  try {
-    const response = await dockerApi.post(
-      `/containers/${sessionId}/${containerId}/stop`,
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "stop Docker container");
-  }
-}
-
-export async function restartDockerContainer(
-  sessionId: string,
-  containerId: string,
-): Promise<{ success: boolean; message: string }> {
-  try {
-    const response = await dockerApi.post(
-      `/containers/${sessionId}/${containerId}/restart`,
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "restart Docker container");
-  }
-}
-
-export async function pauseDockerContainer(
-  sessionId: string,
-  containerId: string,
-): Promise<{ success: boolean; message: string }> {
-  try {
-    const response = await dockerApi.post(
-      `/containers/${sessionId}/${containerId}/pause`,
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "pause Docker container");
-  }
-}
-
-export async function unpauseDockerContainer(
-  sessionId: string,
-  containerId: string,
-): Promise<{ success: boolean; message: string }> {
-  try {
-    const response = await dockerApi.post(
-      `/containers/${sessionId}/${containerId}/unpause`,
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "unpause Docker container");
-  }
-}
-
-export async function removeDockerContainer(
-  sessionId: string,
-  containerId: string,
-  force: boolean = false,
-): Promise<{ success: boolean; message: string }> {
-  try {
-    const response = await dockerApi.delete(
-      `/containers/${sessionId}/${containerId}/remove`,
-      {
-        params: { force },
-      },
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "remove Docker container");
-  }
-}
-
-export async function getContainerLogs(
-  sessionId: string,
-  containerId: string,
-  options?: DockerLogOptions,
-): Promise<{ logs: string }> {
-  try {
-    const response = await dockerApi.get(
-      `/containers/${sessionId}/${containerId}/logs`,
-      {
-        params: options,
-      },
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "get container logs");
-  }
-}
-
-export async function downloadContainerLogs(
-  sessionId: string,
-  containerId: string,
-  options?: DockerLogOptions,
-): Promise<Blob> {
-  try {
-    const response = await dockerApi.get(
-      `/containers/${sessionId}/${containerId}/logs`,
-      {
-        params: { ...options, download: true },
-        responseType: "blob",
-      },
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "download container logs");
-  }
-}
-
-export async function getContainerStats(
-  sessionId: string,
-  containerId: string,
-): Promise<DockerStats> {
-  try {
-    const response = await dockerApi.get(
-      `/containers/${sessionId}/${containerId}/stats`,
-    );
-    return response.data;
-  } catch (error) {
-    throw handleApiError(error, "get container stats");
-  }
-}
+export {
+  connectDockerSession,
+  verifyDockerTOTP,
+  verifyDockerWarpgate,
+  disconnectDockerSession,
+  keepaliveDockerSession,
+  getDockerSessionStatus,
+  validateDockerAvailability,
+  listDockerContainers,
+  getDockerContainerDetails,
+  startDockerContainer,
+  stopDockerContainer,
+  restartDockerContainer,
+  pauseDockerContainer,
+  unpauseDockerContainer,
+  removeDockerContainer,
+  getContainerLogs,
+  downloadContainerLogs,
+  getContainerStats,
+} from "@/api/docker-api";
 
 export interface DashboardLayout {
   cards: Array<{
